Add custom clock instantiation and connection. The user will be allowed to include a new field called 'user_clocks' in the Yaml image core file, and connections to and from it in the 'clk_domains' field. This will look like as follows:
user_clocks:
  uc0:
    ip: user_clock_100
    port_in: &uc0in clk_in1
    port_out: &uc0out clk_out1
...
clk_domains:
- {srcblk: _device_, srcport: radio,   dstblk: uc0,      dstport: *uc0in }
- {srcblk: uc0,     srcport: *uc0out,  dstblk: myBlock, dstport: ce }

Then the image builder will parse this information and instantiate the required module and wires into the generated rfnoc_image_core.v.
The user will be responsible to add the required XCI (created for instance by using the viv_create_ip tool), the XDC constraint file for the new clock, and the Makefile.inc so that the new IP is picked up and built as a bitstream's dependency.